+++ app/src/main/java/org/simple/clinic/patient/PatientLineListRow.kt
@@ -3,7 +3,9 @@ package org.simple.clinic.patient
 import androidx.room.Embedded
 import org.simple.clinic.bp.BloodPressureMeasurement
 import org.simple.clinic.medicalhistory.Answer
+import org.simple.clinic.patient.businessid.BusinessId
 import java.time.Instant
+import java.util.UUID
 
 data class PatientLineListRow(
     val patientName: String,
@@ -12,7 +14,9 @@ data class PatientLineListRow(
     @Embedded
     val age: PatientAgeDetails,
     val registrationDate: Instant,
+    val registrationFacilityId: UUID?,
     val registrationFacilityName: String?,
+    val assignedFacilityId: UUID?,
     val assignedFacilityName: String?,
     val streetAddress: String?,
     val colonyOrVillage: String?,
@@ -20,5 +24,7 @@ data class PatientLineListRow(
     val diagnosedWithHypertension: Answer,
     val diagnosedWithDiabetes: Answer,
     @Embedded(prefix = "bp_")
-    val latestBloodPressureMeasurement: BloodPressureMeasurement?
+    val latestBloodPressureMeasurement: BloodPressureMeasurement?,
+    @Embedded(prefix = "bp_passport_")
+    val bpPassport: BusinessId?
 )